Bug 1336587 - Signatures for crash pings
This is a very simple proof of concept for taking crash ping json and running
it through symbolapi and the Socorro signature generation logic. It relies on an
external symbolapi server, set via the symbolication_api_url conf option.

ping_processor_app takes a list of files each containing one crash ping
on stdin, and produces signatures on stdout.